[compiler] add compound statements to tokens-to-tree.
In the event we get a list of one we can just return it, since we have no tree to deal with.
If we have a list of more than one, attempt to compound the statements until we rerun into a descendant token. If list of compounds greater than 1 wrap in a compound statement then wrap a subtree in descendants. If we do not have a list of compounds greater than one just cons token with the descendants.

Examples:

'((:SYMBOL :DIV)) => '((:SYMBOL :DIV))

'((:SYMBOL :DIV)
   :DESCENDANT (:SYMBOL :P) => '((SYMBOL :DIV)
                                 (:DESCENDANTS ((:SYMBOL :P))))

'((:SYMBOL :DIV) (:ID :MY-ID)) => '((:COMPOUND ((:SYMBOL :DIV)
                                                (:ID :MY-ID))))

'((:SYMBOL :DIV) (:ID :MY-ID)
  :DESCENDANT (:SYMBOL :P)) => '((:COMPOUND ((:SYMBOL :DIV)
                                             (:ID :MY-ID)))
                                 (:DESCENDANTS ((:SYMBOL :P))))

'((:SYMBOL :DIV) (:ID :MY-ID)
  :DESCENDANT (:SYMBOL :P) (:CLASS :MY-CLASS)) => '((:COMPOUND ((:SYMBOL :DIV)
                                                                (:ID :MY-ID)))
                                                    (:DESCENDANTS ((:COMPOUND ((:SYMBOL :P)
                                                                               (:CLASS :MY-CLASS))))))